+++ dio/lib/src/parameter.dart
@@ -1,3 +1,5 @@
+import 'package:collection/collection.dart';
+
 import 'options.dart';
 
 /// Indicates a param being used as queries or form data,
@@ -26,9 +28,12 @@ class ListParam<T> {
       identical(this, other) ||
       other is ListParam &&
           runtimeType == other.runtimeType &&
-          value == other.value &&
+          const DeepCollectionEquality().equals(value, other.value) &&
           format == other.format;
 
   @override
-  int get hashCode => value.hashCode ^ format.hashCode;
+  int get hashCode => Object.hash(
+        const DeepCollectionEquality().hash(value),
+        format,
+      );
 }